@@ -2,19 +2,20 @@ ext {
 
     android = [
             applicationId    : 'com.example.yxm.photogenic',
-            compileSdkVersion: 28,
+            compileSdkVersion: 29,
             minSdkVersion    : 19,
-            targetSdkVersion : 28,
-            versionCode      : 1,
-            versionName      : "1.0",
+            buildToolsVersion: "29.0.0",
+            targetSdkVersion : 29,
+            versionCode      : 2,
+            versionName      : "1.1",
             mutidex          : true,
             multiDexEnabled  : true
     ]
 
     dependLibsVersion = [
-            appcompat               : '28.0.0',
-            design                  : '28.0.0',
-            recyclerview            : '28.0.0',
+            appcompat               : '1.0.0',
+            design                  : '1.0.0',
+            recyclerview            : '1.0.0',
             retrofit2               : "2.4.0",
             rxjava2                 : "2.2.9",
             glide                   : "4.9.0",
