@@ -1,8 +1,7 @@
 package com.example.yxm.photogenic.module.categorydetails
 
-import android.widget.ImageView
 import com.chad.library.adapter.base.BaseQuickAdapter
-import com.chad.library.adapter.base.BaseViewHolder
+import com.chad.library.adapter.base.viewholder.BaseViewHolder
 import com.example.lib_imageloader.ImageLoaderManager
 import com.example.lib_network.bean.CategoriesBean
 import com.example.yxm.photogenic.R
@@ -13,8 +12,8 @@ import com.example.yxm.photogenic.R
  */
 class CategoryAllAdapter: BaseQuickAdapter<CategoriesBean, BaseViewHolder>(R.layout.item_big_category) {
 
-    override fun convert(helper: BaseViewHolder?, item: CategoriesBean) {
-        helper?.run {
+    override fun convert(holder: BaseViewHolder, item: CategoriesBean) {
+        holder.run {
             ImageLoaderManager.displayImageForView(getView(R.id.category_cover_image),item.bgPicture)
             setText(R.id.category_name,"#${item.name}")
         }
