@@ -1,7 +1,7 @@
 package com.example.yxm.photogenic.module.categorydetails
 
 import com.chad.library.adapter.base.BaseQuickAdapter
-import com.chad.library.adapter.base.BaseViewHolder
+import com.chad.library.adapter.base.viewholder.BaseViewHolder
 import com.example.lib_network.bean.CommonVideoBean
 import com.example.yxm.photogenic.R
 
@@ -9,11 +9,11 @@ import com.example.yxm.photogenic.R
  *Created by yxm on 2020/2/10
  *@function  视频标签adapter
  */
-class TagAdapter: BaseQuickAdapter<CommonVideoBean.ResultBean.ResultData.TagBean,BaseViewHolder>(R.layout.item_tag) {
+class TagAdapter : BaseQuickAdapter<CommonVideoBean.ResultBean.ResultData.TagBean, BaseViewHolder>(R.layout.item_tag) {
 
-    override fun convert(helper: BaseViewHolder, item: CommonVideoBean.ResultBean.ResultData.TagBean) {
-        with(helper){
-            setText(R.id.tag_tv,item.name)
+    override fun convert(holder: BaseViewHolder, item: CommonVideoBean.ResultBean.ResultData.TagBean) {
+        with(holder) {
+            setText(R.id.tag_tv, item.name)
         }
     }
 }
