@@ -2,14 +2,13 @@ package com.example.yxm.photogenic.module.community
 
 import android.widget.FrameLayout
 import android.widget.ImageView
-import com.bumptech.glide.Glide
 import com.chad.library.adapter.base.BaseQuickAdapter
-import com.chad.library.adapter.base.BaseViewHolder
+import com.chad.library.adapter.base.viewholder.BaseViewHolder
 import com.example.lib_imageloader.ImageLoaderManager
 import com.example.lib_network.bean.CommunityBean
 import com.example.yxm.photogenic.R
 import com.example.yxm.photogenic.utils.ScreenHelper
-import de.hdodenhof.circleimageview.CircleImageView
+import com.example.yxm.photogenic.widget.CircleImageView
 
 /**
  *Created by yxm on 2020/2/20
@@ -17,10 +16,10 @@ import de.hdodenhof.circleimageview.CircleImageView
  */
 class CommunityRecommendAdapter : BaseQuickAdapter<CommunityBean.Issue, BaseViewHolder>(R.layout.item_community_recommend) {
 
-    override fun convert(helper: BaseViewHolder, item: CommunityBean.Issue) {
+    override fun convert(holder: BaseViewHolder, item: CommunityBean.Issue) {
         when (item.type) {
-            "pictureFollowCard" -> initPictureItem(helper, item)
-            "autoPlayFollowCard" -> initVideoItem(helper, item)
+            "pictureFollowCard" -> initPictureItem(holder, item)
+            "autoPlayFollowCard" -> initVideoItem(holder, item)
         }
     }
 
@@ -30,21 +29,20 @@ class CommunityRecommendAdapter : BaseQuickAdapter<CommunityBean.Issue, BaseView
     private fun initPictureItem(helper: BaseViewHolder, item: CommunityBean.Issue) {
         with(helper) {
             if (item.data.content.data.urls!!.size <= 1) {
-                setGone(R.id.type_label, false)
+                setVisible(R.id.type_label, false)
             } else {
-                setGone(R.id.type_label, true)
+                setVisible(R.id.type_label, true)
                 setImageResource(R.id.type_label, R.drawable.abc_ic_menu_copy_mtrl_am_alpha)
             }
             displayScaleImage(getView<ImageView>(R.id.community_cover_image).layoutParams as FrameLayout.LayoutParams,
                     getView(R.id.community_cover_image), item.data.content.data.cover.feed ?: "",
                     item.data.content.data.width.toInt(), item.data.content.data.height.toInt())
-            Glide.with(getView<CircleImageView>(R.id.author_avatar).context)
-                    .load(item.data.header.icon)
-                    .into(getView<CircleImageView>(R.id.author_avatar))
+
+            ImageLoaderManager.displayImageForView(getView<CircleImageView>(R.id.author_avatar) as ImageView,item.data.header.icon)
             setText(R.id.picture_brief_introduce, item.data.content.data.description)
             setText(R.id.author_name, item.data.header.issuerName)
             setText(R.id.like_count, "${item.data.content.data.consumption.collectionCount}")
-            addOnClickListener(R.id.like_iv)
+            addChildClickViewIds(R.id.like_iv)
         }
     }
 
@@ -53,18 +51,16 @@ class CommunityRecommendAdapter : BaseQuickAdapter<CommunityBean.Issue, BaseView
      */
     private fun initVideoItem(helper: BaseViewHolder, item: CommunityBean.Issue) {
         with(helper) {
-            setGone(R.id.type_label, true)
+            setVisible(R.id.type_label, true)
             setImageResource(R.id.type_label, R.drawable.ic_play_circle_outline_white_48dp)
             displayScaleImage(getView<ImageView>(R.id.community_cover_image).layoutParams as FrameLayout.LayoutParams,
                     getView(R.id.community_cover_image), item.data.content.data.cover.feed ?: "",
                     item.data.content.data.width.toInt(), item.data.content.data.height.toInt())
-            Glide.with(getView<CircleImageView>(R.id.author_avatar).context)
-                    .load(item.data.header.icon)
-                    .into(getView<CircleImageView>(R.id.author_avatar))
+            ImageLoaderManager.displayImageForView(getView<CircleImageView>(R.id.author_avatar) as ImageView,item.data.header.icon)
             setText(R.id.picture_brief_introduce, item.data.content.data.description)
             setText(R.id.author_name, item.data.header.issuerName)
             setText(R.id.like_count, "${item.data.content.data.consumption.collectionCount}")
-            addOnClickListener(R.id.like_iv)
+            addChildClickViewIds(R.id.like_iv)
         }
     }
 
