@@ -2,7 +2,7 @@ package com.example.yxm.photogenic.module.searchresult
 
 import com.bumptech.glide.Glide
 import com.chad.library.adapter.base.BaseQuickAdapter
-import com.chad.library.adapter.base.BaseViewHolder
+import com.chad.library.adapter.base.viewholder.BaseViewHolder
 import com.example.lib_imageloader.ImageLoaderManager
 import com.example.lib_network.bean.CommonVideoBean
 import com.example.yxm.photogenic.R
@@ -15,8 +15,8 @@ import de.hdodenhof.circleimageview.CircleImageView
  */
 class SearchResultAdapter : BaseQuickAdapter<CommonVideoBean.ResultBean, BaseViewHolder>(R.layout.item_big_video_bean) {
 
-    override fun convert(helper: BaseViewHolder, item: CommonVideoBean.ResultBean) {
-        with(helper) {
+    override fun convert(holder: BaseViewHolder, item: CommonVideoBean.ResultBean) {
+        with(holder) {
             ImageLoaderManager.displayImageForView(getView(R.id.video_picture), item.data.cover.feed
                     ?: "")
             Glide.with((getView(R.id.video_author_avatar) as CircleImageView).context)
@@ -25,7 +25,7 @@ class SearchResultAdapter : BaseQuickAdapter<CommonVideoBean.ResultBean, BaseVie
             setText(R.id.video_title, item.data.title)
             setText(R.id.video_duration, TimeHelper.secToTime(item.data.duration))
             setText(R.id.video_secondary_title, "#${item.data.category}")
-            addOnClickListener(R.id.video_share_iv)
+            addChildClickViewIds(R.id.video_share_iv)
         }
     }
 }
