@@ -2,26 +2,25 @@ package com.example.yxm.photogenic.ui.activity
 
 import android.content.Intent
 import android.os.Bundle
-import android.support.design.widget.AppBarLayout
-import android.support.design.widget.CollapsingToolbarLayout
-import android.support.v4.widget.NestedScrollView
-import android.support.v7.widget.LinearLayoutManager
-import android.support.v7.widget.RecyclerView
-import android.support.v7.widget.Toolbar
 import android.view.View
 import android.widget.ImageView
 import android.widget.TextView
+import androidx.appcompat.widget.Toolbar
+import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
 import com.chad.library.adapter.base.BaseQuickAdapter
 import com.example.lib_imageloader.ImageLoaderManager
 import com.example.lib_network.bean.CategoriesBean
 import com.example.lib_network.bean.CategoryDetailBean
-import com.example.lib_share.share.ShareManager
+import share.core.ShareManager
 import com.example.yxm.photogenic.R
 import com.example.yxm.photogenic.base.BaseActivity
 import com.example.yxm.photogenic.font.FontTextView
 import com.example.yxm.photogenic.module.categorydetails.*
 import com.example.yxm.photogenic.utils.AppBarStateChangeListener
 import com.example.yxm.photogenic.widget.FooterView
+import com.google.android.material.appbar.AppBarLayout
+import com.google.android.material.appbar.CollapsingToolbarLayout
 import com.gyf.immersionbar.ktx.immersionBar
 import com.scwang.smart.refresh.header.ClassicsHeader
 import com.scwang.smart.refresh.layout.SmartRefreshLayout
@@ -112,7 +111,6 @@ class CategoryDetailActivity : BaseActivity(), TagDetailContract.ITagDetailView 
                 }
             }
         })
-
         //appbar滑动状态监听
         mAppbar.addOnOffsetChangedListener(object : AppBarStateChangeListener() {
             override fun onStateChanged(appBarLayout: AppBarLayout, state: State) {
@@ -150,7 +148,7 @@ class CategoryDetailActivity : BaseActivity(), TagDetailContract.ITagDetailView 
             mCategoryDetailPresenter.getTagsVideo(categoryBean.tagId.toLong())
         }
 
-        mAdapter.onItemClickListener = BaseQuickAdapter.OnItemClickListener { adapter, _, position ->
+        mAdapter.setOnItemClickListener { adapter, _, position ->
             val item = adapter.getItem(position) as CategoryDetailBean.FollowCardBean
             val bundle = Bundle().apply {
                 putSerializable("video", item.data.content.data)
@@ -162,7 +160,7 @@ class CategoryDetailActivity : BaseActivity(), TagDetailContract.ITagDetailView 
             })
         }
         //分享
-        mAdapter.onItemChildClickListener = BaseQuickAdapter.OnItemChildClickListener { adapter, _, position ->
+        mAdapter.setOnItemChildClickListener{ adapter, _, position ->
             val item = adapter.getItem(position) as CategoryDetailBean.FollowCardBean
             ShareManager.shareWebPage(mContext,
                     item.data.content.data.description,
