@@ -2,12 +2,12 @@ package com.example.yxm.photogenic.ui.fragment
 
 import android.content.Intent
 import android.os.Bundle
-import android.support.v7.widget.LinearLayoutManager
-import android.support.v7.widget.RecyclerView
 import android.view.View
+import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
 import com.chad.library.adapter.base.BaseQuickAdapter
 import com.example.lib_network.bean.HomeBean
-import com.example.lib_share.share.ShareManager
+import share.core.ShareManager
 import com.example.yxm.photogenic.R
 import com.example.yxm.photogenic.base.BaseFragment
 import com.example.yxm.photogenic.module.home.HomeReportAdapter
@@ -87,7 +87,7 @@ class HomePageDailyReportFragment : BaseFragment(), HomeReportContract.IHomeRepo
             adapter = mAdapter
         }
 
-        mAdapter.onItemClickListener = BaseQuickAdapter.OnItemClickListener { adapter, _, position ->
+        mAdapter.setOnItemClickListener{ adapter, _, position ->
             val type = adapter.getItemViewType(position)
             val item = adapter.getItem(position) as HomeBean.Issue
             val bundle = Bundle().apply {
@@ -104,7 +104,7 @@ class HomePageDailyReportFragment : BaseFragment(), HomeReportContract.IHomeRepo
             }
         }
 
-        mAdapter.onItemChildClickListener = BaseQuickAdapter.OnItemChildClickListener { adapter, view, position ->
+        mAdapter.setOnItemChildClickListener { adapter, view, position ->
             val item = adapter.getItem(position) as HomeBean.Issue
             //分享区别数据类型
             if (view.id == R.id.video_share_iv) {
