@@ -1,8 +1,8 @@
 package com.example.yxm.photogenic.ui.fragment
 
-import android.support.v7.widget.LinearLayoutManager
-import android.support.v7.widget.RecyclerView
 import android.view.View
+import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
 import com.chad.library.adapter.base.BaseQuickAdapter
 import com.example.yxm.photogenic.R
 import com.example.yxm.photogenic.base.BaseFragment
@@ -18,7 +18,7 @@ import org.greenrobot.eventbus.EventBus
  * Created by yxm on 2020-1-16
  * @function: 热搜词汇fragment
  */
-class HotWordsFragment: BaseFragment(), HotWordsContract.IHotWordsView{
+class HotWordsFragment : BaseFragment(), HotWordsContract.IHotWordsView {
 
     private lateinit var hotWordsRv: RecyclerView
 
@@ -66,7 +66,7 @@ class HotWordsFragment: BaseFragment(), HotWordsContract.IHotWordsView{
     }
 
     override fun initListener() {
-        mAdapter.onItemClickListener = BaseQuickAdapter.OnItemClickListener { adapter, _, position ->
+        mAdapter.setOnItemClickListener { adapter, _, position ->
             val item = adapter.getItem(position) as String
             KeyBoardHelper.hideKeyBoard(hotWordsRv)
             EventBus.getDefault().post(HotWordsEvent(item))
