@@ -2,12 +2,12 @@ package com.example.yxm.photogenic.ui.fragment
 
 import android.content.Intent
 import android.os.Bundle
-import android.support.v7.widget.LinearLayoutManager
-import android.support.v7.widget.RecyclerView
 import android.view.View
+import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
 import com.chad.library.adapter.base.BaseQuickAdapter
 import com.example.lib_network.bean.CommonVideoBean
-import com.example.lib_share.share.ShareManager
+import share.core.ShareManager
 import com.example.yxm.photogenic.R
 import com.example.yxm.photogenic.base.BaseFragment
 import com.example.yxm.photogenic.module.categorydetails.RelativeVideoAdapter
@@ -86,20 +86,8 @@ class RankFragment : BaseFragment(), RankContract.IRankView {
         mRefreshLayout.setOnRefreshListener {
             mRankPresenter.getRankData(strategy)
         }
-
-        mAdapter.onItemClickListener = BaseQuickAdapter.OnItemClickListener { adapter, _, position ->
-            val bean = adapter.getItem(position) as CommonVideoBean.ResultBean
-            startActivity(Intent(mContext, VideoPlayActivity::class.java).apply {
-                val bundle = Bundle()
-                bundle.putSerializable("video", bean)
-                bundle.putLong("relativeVideoId", bean.data.id)
-                bundle.putInt("fromWhere", RANK_FRAGMENT)
-                putExtras(bundle)
-            })
-        }
-
         //分享
-        mAdapter.onItemChildClickListener = BaseQuickAdapter.OnItemChildClickListener { adapter, view, position ->
+        mAdapter.setOnItemChildClickListener { adapter, view, position ->
             val item = adapter.getItem(position) as CommonVideoBean.ResultBean
             if (view.id == R.id.video_share_iv) {
                 ShareManager.shareWebPage(mContext,
@@ -108,6 +96,17 @@ class RankFragment : BaseFragment(), RankContract.IRankView {
                         item.data.cover.feed ?: "",
                         item.data.webUrl.raw)
             }
+        }
+
+        mAdapter.setOnItemClickListener { adapter, view, position ->
+            val bean = adapter.getItem(position) as CommonVideoBean.ResultBean
+            startActivity(Intent(mContext, VideoPlayActivity::class.java).apply {
+                val bundle = Bundle()
+                bundle.putSerializable("video", bean)
+                bundle.putLong("relativeVideoId", bean.data.id)
+                bundle.putInt("fromWhere", RANK_FRAGMENT)
+                putExtras(bundle)
+            })
         }
     }
 
