@@ -1,8 +1,8 @@
 package com.example.yxm.photogenic.widget
 
 import android.content.Context
-import android.support.v7.widget.AppCompatTextView
 import android.util.AttributeSet
+import androidx.appcompat.widget.AppCompatTextView
 import java.util.*
 
 /**
