@@ -1,35 +1,34 @@
 package share.adapter
 
 import com.chad.library.adapter.base.BaseQuickAdapter
-import com.chad.library.adapter.base.BaseViewHolder
+import com.chad.library.adapter.base.viewholder.BaseViewHolder
 import com.example.lib_share.R
 
 /**
  *Created by yxm on 2020/2/22
  *@function 分享平台adapter
  */
-class SharePlatformAdapter: BaseQuickAdapter<PlatformBean,BaseViewHolder>(R.layout.item_share_platform){
+class SharePlatformAdapter: BaseQuickAdapter<PlatformBean, BaseViewHolder>(R.layout.item_share_platform){
 
-    override fun convert(helper: BaseViewHolder, item: PlatformBean) {
-        with(helper){
+    override fun convert(holder: BaseViewHolder, item: PlatformBean) {
+        with(holder){
             when(item.name){
                 "QQ" -> {
-                    setBackgroundRes(R.id.platform_icon,R.drawable.bg_share_qq)
+                    setBackgroundResource(R.id.platform_icon,R.drawable.bg_share_qq)
                 }
                 "QQ空间" -> {
-                    setBackgroundRes(R.id.platform_icon,R.drawable.bg_share_qzone)
+                    setBackgroundResource(R.id.platform_icon,R.drawable.bg_share_qzone)
                 }
                 "微信" -> {
-                    setBackgroundRes(R.id.platform_icon,R.drawable.bg_share_friend)
+                    setBackgroundResource(R.id.platform_icon,R.drawable.bg_share_friend)
                 }
                 "朋友圈" -> {
-                    setBackgroundRes(R.id.platform_icon,R.drawable.bg_share_moment)
+                    setBackgroundResource(R.id.platform_icon,R.drawable.bg_share_moment)
                 }
                 "微博" -> {
-                    setBackgroundRes(R.id.platform_icon,R.drawable.bg_share_weibo)
+                    setBackgroundResource(R.id.platform_icon,R.drawable.bg_share_weibo)
                 }
             }
-            //setImageResource(R.id.platform_icon,item.iconId)
             setText(R.id.platform_name,item.name)
         }
     }
